+++ CHANGELOG.md
@@ -9,8 +9,8 @@ and this project (loosely) adheres to [Semantic Versioning](https://semver.org/s
 
 ### Changed
 
-- The default `typ` header for issued SD-JWT VCs is now `dc+sd-jwt`.
-  - The `Verifier` will accept both `vc+sd-jwt` and `dc+sd-jwt`.
+- The `typ` header for issued SD-JWT VCs is now `dc+sd-jwt` as per [draft-08](https://www.ietf.org/archive/id/draft-ietf-oauth-sd-jwt-vc-08.html#section-3.2.1)
+- The `Verifier` will accept both `vc+sd-jwt` and `dc+sd-jwt`.
 
 ## 1.3.0 - 2024-10-07